fixed bug in blur_gauss() with latest version of scipy.signal.fftconvolve

+++ ehtim/image.py
@@ -1395,7 +1395,7 @@ class Image(object):
             return imarr_blur
 
         # Convolve the primary image
-        imarr = (self.imvec).reshape(self.ydim, self.xdim)
+        imarr = (self.imvec).reshape(self.ydim, self.xdim).astype('float64')
         imarr_blur = blur(imarr, gauss)
 
         # Make new image object
@@ -1409,7 +1409,7 @@ class Image(object):
                 continue
             polvec = self._imdict[pol]
             if len(polvec):
-                polarr = polvec.reshape(self.ydim, self.xdim)
+                polarr = polvec.reshape(self.ydim, self.xdim).astype('float64')
                 if frac_pol:
                     polarr = blur(polarr, gausspol)
                 outim.add_pol_image(polarr, pol)
@@ -1418,7 +1418,7 @@ class Image(object):
         mflist_out = []
         for mfvec in self._mflist:
             if len(mfvec):
-                mfarr = mfvec.reshape(self.ydim, self.xdim)
+                mfarr = mfvec.reshape(self.ydim, self.xdim).astype('float64')
                 mfarr = blur(mfarr, gauss)
                 mfvec_out = mfarr.flatten()
             else:
